V1.0 Release:
Finished Functions:
	reslv (nres pres)
	atp (kb a)
	occurs? (var x)
	subs (clause +theta+)

Modified subs and occurs? to correspond to code discussed in class.
Finished reslv to check for loops, print results, and test for the
null clause.

V1.1 Plans:

Modify a few functions for ease-of-use.
Improve upon the get-next-res function.
Change the way resolution works.
Maybe change the struct compound to a list.